implement level-specific paintings (BOXPAINT)
you can now define a BOXPAINT lump to add in new paintings into box party.

this lump allows you to control which sprite to load, the chance of it spawning, which level it can spawn in (or optionally spawn in every level), and also the size of the painting.

many of the already existing paintings have been marked to only spawn in specific levels now. this does decrease the variety of paintings in each level, but that can be mitigated by making more paintings.

you can also define a painting multiple times if you want it to spawn in multiple specific levels.

does this have any impact on gameplay? no.

was it worth spending time and effort coding? yes.